+++ src/neo/SmartContract/Callbacks/CallbackBase.cs
@@ -1,4 +1,4 @@
-namespace Neo.SmartContract
+namespace Neo.SmartContract.Callbacks
 {
     internal abstract class CallbackBase
     {
